refactor: 스케줄러 이름 변경
